app/requirements-prod.txt
# Production-optimized requirements for Oxtari
# Core dependencies
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.5.0,<3.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database and vector store
weaviate-client>=3.25.0,<4.0.0
psycopg2-binary>=2.9.0,<3.0.0

# HTTP clients
httpx>=0.25.0,<0.26.0
aiohttp>=3.9.0,<4.0.0
requests>=2.31.0,<3.0.0

# AI/ML libraries
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
sentence-transformers>=2.2.0,<3.0.0

# Document processing
pypdf>=3.17.0,<4.0.0
python-docx>=1.1.0,<2.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<5.0.0

# Utilities
wasabi>=1.1.0,<2.0.0
tqdm>=4.66.0,<5.0.0
python-json-logger>=2.0.0,<3.0.0

# Security
cryptography>=41.0.0,<42.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Optional: Only include if needed
# openai>=1.3.0,<2.0.0
# anthropic>=0.7.0,<1.0.0
# cohere>=4.37.0,<5.0.0
# groq>=0.4.0,<1.0.0